refactor(windows): move uninstall internals into a dedicated Windows package
surfaces/cli/lifecycle/uninstall.py had grown to 1120 lines, most of it Windows
process discovery, layout classification, locking, scheduling, and a 460-line
PowerShell worker embedded as a Python string literal.

Windows behavior now lives in surfaces/cli/lifecycle/windows/ (layout,
processes, powershell, cleanup), and uninstall.py keeps only cross-platform
orchestration at 270 lines. The cleanup worker is a real uninstall_cleanup.ps1
package resource: shipped through setuptools package-data and PyInstaller's
surfaces.cli data collection, still copied to a private temp file before the
detached run, and now parsed directly by the PowerShell AST in a test.

The legacy_executable field was dead on every path, so it is gone along with
the extra_paths branch it fed. The second classification in run_uninstall is
deliberate revalidation after the confirmation prompt and is kept, now with a
comment saying so.

No security, ownership, rollback, or cleanup behavior changes.
