+++ pyproject.toml
@@ -165,6 +165,7 @@ include = [
 ]
 
 [tool.setuptools.package-data]
+"surfaces.cli.lifecycle.windows" = ["*.ps1"]
 "surfaces.shared" = ["sample_alerts/*.json"]
 "core.agent_harness.prompts" = ["opensre_system_prompt.md", "skills/**/*.md"]
 "integrations" = ["**/SKILL.md", "yandex_cloud/api_index.json"]
